+++ tests/test_common.py
@@ -9,7 +9,7 @@ from .utils import ContextManagerDBTesting
 @pytest.mark.skipif(SERVICE_UNAVAILABLE, reason="Database is not available")
 def test_get_stats(bbagent_obj):
     with ContextManagerDBTesting(config=bbagent_obj.config, add_data=True, bedset=True):
-        return_result = bbagent_obj.get_stats
+        return_result = bbagent_obj.get_stats()
 
         assert return_result
         assert return_result.bedfiles_number == 1