+++ docs/changelog.md
@@ -2,6 +2,12 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) and [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format.
 
+# [0.7.0] - 2024-09-20
+## Added 
+- Table and methods for reference genome validator
+- Table with standard metadata schema
+- Bed file opening improvements
+
 # [0.6.1] - 2024-08-21
 ## Added